warranty != liability
warranty = you get something and it's guaranteed to work and not fail within a certain period of time let's say a laptop has a 3 year warranty: you get a free laptop if for example the USB port breaks within the first 10 days

liability = something else. Basically who's responsible (and will probably pay up) for damages caused to something. You rear end a car on purpose, you're gonna be liable for all the damages.

Reminder that in `Software` repository, we will not be liable for any damages caused by any material present in this repository, and any material comes with no warranty (the script doesn't even have to work but we try our best of course)

Good day folks!

Signed-off-by: GmaerSoft42 <162634737+GmaerSoft42@users.noreply.github.com>